refactored tests paths
